requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
pickle-mixin==1.0.2
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.5.0
tzdata==2025.1
Werkzeug==3.1.3
wheel==0.45.1
xgboost==2.1.4
